test(browser): move ARIA-budget cases to dedicated aria-compressor.test.ts
Split the summariseAriaSnapshot suite out of browser-tools.test.ts (617->487
lines, under the 300+ file-size concern) into aria-compressor.test.ts. Also
drops the stray pnpm-lock.yaml that got committed from a local pnpm install;
repo uses npm/package-lock.json.
